Add script to set the name of the web server
> Usage: bin/set-web-server-name <hostname|fqdn>

First argument is a host name or a fully qualified domain name.
This name must be resolved to the correct IP/address for every
client that visits the mirror MusicBrainz web server on the VM.

Default name is 'localhost' so that the web server can be visited
from the VM host.  This script is useful to allow visiting the web
server from other hosts than the VM host (if name can be resolved).